feat: allow float in report target api

+++ owehub/shared/models/getReportsTargetModels.go
@@ -85,11 +85,11 @@ type ProductionTargetsByYearRespItem struct {
 type UpdateProductionTargetsReqItem struct {
 	Year         int     `json:"year"`
 	Month        int     `json:"month"`
-	ProjectsSold int64   `json:"projects_sold"`
+	ProjectsSold float64 `json:"projects_sold"`
 	MwSold       float64 `json:"mw_sold"`
-	InstallCt    int64   `json:"install_ct"`
+	InstallCt    float64 `json:"install_ct"`
 	MwInstalled  float64 `json:"mw_installed"`
-	BatteriesCt  int64   `json:"batteries_ct"`
+	BatteriesCt  float64 `json:"batteries_ct"`
 }
 
 type UpdateProductionTargetsReq struct {
